Clean up tableExists check
The postgres output plugin would check if a required table
was present in a local cache before writing data to it.
If it was missing it would issue a "CREATE TABLE ..." statement.
The cache was maintained in a simple map and was accessed
from anywhere in the plugin code with no consistent way how it was
updated or queried. This commit introduces an abstracted Table Keeper
which handles all operations surrounding the cache